Implement heap for priority.
Added 2 optional parameters: priority and metadata
By default, priority=1.
The higher the priority, the faster the request will be processed. This can be useful in cases where you receive many requests simultaneously on model-hub and need to ensure faster access to workers for more important tasks.
metadata is useful in cases where it's necessary to understand from the logs where and how a request is being processed. Metadata can be a string with any content. You will see this string in the logs after the prediction has been successfully made.